requirements.txt
colorama==0.4.4
pydub==0.25.1
python-dotenv==0.21.0
supabase==0.7.1
SpeechRecognition==3.8.1
numpy==1.23.5